move and rename proxy interfaces
