+++ src/main/java/com/wine/to/up/winestyle/parser/service/service/UnstableLoader.java
@@ -1,0 +1,5 @@
+package com.wine.to.up.winestyle.parser.service.service;
+
+public interface UnstableLoader extends WebPageLoader {
+    int getFailuresCount();
+}
