+++ src/main/java/com/wine/to/up/winestyle/parser/service/service/WebPageLoader.java
@@ -1,9 +1,9 @@
-package com.wine.to.up.winestyle.parser.service.service.implementation.document;
+package com.wine.to.up.winestyle.parser.service.service;
 
 import org.jsoup.nodes.Document;
 
 import java.io.IOException;
 
-public interface IWebPageLoader {
+public interface WebPageLoader {
     Document getDocument(String url) throws IOException;
 }
