--- src/main/java/com/wine/to/up/winestyle/parser/service/service/implementation/document/IUnstableLoader.java
@@ -1,5 +1,0 @@
-package com.wine.to.up.winestyle.parser.service.service.implementation.document;
-
-public interface IUnstableLoader extends IWebPageLoader {
-    int getFailuresCount();
-}
